refactor: split logic of data digestion from DataModel Pojo
preparation for unit testing the digestion
